api-server/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1

# AWS Bedrock
boto3==1.36.7
botocore==1.36.7

# Monitoring
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.2

# Security
cryptography==44.0.0

# MFA / TOTP
pyotp==2.9.0

# WebSocket
websockets==14.2

# SSH proxy
asyncssh==2.18.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
